Fix identification of templates from configuration
To identify templates (as header, row, or CommonsCat),
each template of the page is compared to the one recorded
in the configuration.

But the template as returned is fully qualified with the Namespace
(eg Page(Modèle:Ligne de tableau Monuments (Suisse))).
To get its actual name, we need to use the title method

Change-Id: Idca0ba57be703a52fcad258e5d79d6fb62946265