Support for Wikidata while guessing categories from CommonsCat
Wikipedia lists may have an empty CommonsCat template,
not because the target is the title of the page itself
(which is currently assumed by the bot), but because
the CommonsCat is taken from Wikidata.

This changeset adds support for this use case to ErfgoedBot.

If the bot finds an empty CommonsCat, it will try to access the
linked Wikidata item, and try to extract the target of P373.

Bug: T110003
Change-Id: Ie5b66d175a03c8e325a86d6eddb607057dca1307