Rename processText to processPage and remove processTextfile
The text parameter of processText is never used. This method always
gets the contents from the given page. It thus makes more sense
to call it processPage

processTextfile was just a proxy method for processText,
handing to processText the contents of a text file.
Since processText just ignores that anyway, remove that
method as it is patently unused.

Change-Id: Ia835a1e7c9aca213fe3808fa29f1dffffa0d2b79