Add test for checkLat/checkLon and test checkers/converters are triggered
This is in preparation of the re-factoring done in I28895b6cc94e87af68f88b455ef2a431d4b5f00a .

The longitude tests are almost identical to the latitude tests since the
two calls don't share any code (yet). It should be possible to do this in
a better way but I'm unsure how this would trigger setUp() tearDown() calls.

Note that this includes the test_trigger_problematic_check which is
expected to fail once the re-factoring is done.

Also:
* Make TestLookupSourceField use differently named source/dest values
  for clarity. And as a result changed function name.
* Change all assert_called_with to assert_called_once_with to make it
  explicit that only a single call is expected.
* Add test for lat_no_lon and lon_no_lat warnings.

Change-Id: I8f48c744ee7ce5d57ddce7f12e18d5ada41c75ce