Ensure all monuments_config entries are used
Having unused entries makes it hard to spot changes to fill_monuments_all
which orphan datasets as well as making triaging of monumnets_config even
more cumbersome.

- Remove all unused entries in monuments_config (where table name starts
with "monuments").

- Add test ensuring that entries in monuments_config (where table name starts
with "monuments") are also used in fill_monuments_all.

Ideally the same test should be done for wlpa AND it should be ensured
that all entries are either in one or the other.

Bug: T136704
Change-Id: I619230e8ce75f87c3606d1be0dcdc1d8f3a06c9e