Remove illegal uses of generateRegistrantUrl
The generateRegistrantUrl converter needs to have a
registrantUrlBase defined in the config to work.

Since d2c242a, this throws an Exception for 16 country configs:
Un-defined converter in config for monuments_xx_(xx): generateRegistrantUrl

- Remove all illegal uses of generateRegistrantUrl

- Add test ensuring that country configs using the generateRegistrantUrl
converter also has a non-empty registrantUrlBase, and that this URL
contains a '%s' for substitution with the identifier.

Bug: T137096
Change-Id: I3c964a7c7ee8feedec005612dc1cc891bdd43682