Centralize methods to connect to database in one file
The methods to connect to the databases (monuments and commons)
were repeated in every single file.

This centralizes in one for easier maintenance and reduce dependencies.

It also paves the way for using dependency injection
to avoid a hard dependency on MySQL.

Also rename the methods to clarify their purposes.

Change-Id: I10c6d1733b0a1c02ea3eaeaaab93dac4fb8cf19d