Shell script for extracting the last full log
A script which can be run from anyware on toollabs to extract the
log from the last complete run. Useful for isolating errors etc.

Note that this will not catch the logs for runs which fail so
fatally that the update_monuments.sh script fails to output its
final log line.

Change-Id: I15e96ff856a954bbdcb44529632d1f2edcdfc455