Mount SQL files via a volume instead of copying in Docker setup
Copying SQL files means that for new files to be taken into account,
one has to rebuild the image.
This makes sense if these are fixtures that rarely change,
but for erfgoedbot, creating new SQL fixtures files is part of
the normal development process for adding new countries.

Also expand ReadMe on development environment,
with step-by-step explanation on how to run harvesting
(which is likely the most common development goal).

Change-Id: Ib7a7aa56c15444c59c7299b42c0b38944b2a817e