Extract method normalize_identifier and add unit tests
This extracts the method normalize_identifier from processSource,
and add some unit tests for its current behaviour.

It also disentangles the decoding of the identifier
and that of the page name (which may be the actual
reason for the error that we see in the logs).

It keeps the current behaviour of not tracking
images where the title fails to be decoded,
in the spirit of not breaking functionality
in the refactoring steps (although unit tests
have not yet reproduced that behaviour).

Change-Id: I407ce29dc0ea14cc3bcfe617ad4bedad3bf58aca