Ensure sql transactions are committed
Our upgrade in db seems to now allow for transactions. As a result
the transactions need to be committed in order for the results to
actually get stored back in the database.

@Note:
The commons database connection have not been closed in the same
way since these are never used for writing.

Change-Id: I6b60422303ec0c0679530539df881e210d04e351