Harvest the source page of unknown fields
The UnknownFields report is as of now not actionable as it is very
hard to track down the location of the error.

For each field, we store a counter mapping a page and the occurence count.
We display a subset of these as wikilinks on the report in a third column.

Bug: T117330
Change-Id: If1144827c69bbd271aff93a9ec81164b65b29b93