Respect `skip` config when making stats
We were hitting a TypeError during `missing_commonscat_links`
and `unused_monument_images` run:
```
unsupported operand type(s) for +=: 'int' and 'NoneType'
```

This happened when processing `('nl-wd', 'nl')`,
as it was skipped in the processing (per 69327900)
and no totals was computed for it. However, `makeStatistics`
was re-iterating over the entire monuments config.

Bug: T176528
Change-Id: I7de794be6381d2527c6288b546dd2927ce8dc563