Replace database connection library MysqlDB with PyMySQL
`MySQL-python` is a Python interface to MySQL that had not
been udpated in years. It also needs MySql installed locally,
typically resulting when trying to run locally with:
```
ImportError: libmysqlclient.so.18: cannot open shared object file: No such file or directory
```

`PyMySQL` is a pure-Python MySQL client library,
designed as a a drop-in replacement for `MySQLdb`.

See https://github.com/PyMySQL/PyMySQL

Change-Id: Iea3dfb6d71c196aceddb321b599b62bd6b5bec4c