Fix bad idx entries
The idx field in the Statistics table was an ugly hack which
concatenated four values only to later explode them. But if the
strings were too long for the field parts were left out and as
a result exploding the value returned the wrong number of fields.

Aditionally something unidentified edited the values in transit
preventing controlled croping of the values pre concatenation.

The solution is to replace the concatenated field with four explicit
fields in the table.

The one known regression is that it is no longer possible to filter
deeper than on country. Since this possibility was not documented and
the filter url parameter was called stcountry this is unlikely to
cause any major issues.

Bug: T174503
Change-Id: I42f6b0d97f52e99ac79a3960c21e91c1721b600c