Use new database replicas for accessing Commons table
The WMF Cloud Services team now recommends using
a different database cluster for wiki replicas.

See announcement at J70:
https://phabricator.wikimedia.org/phame/post/view/70/new_wiki_replica_servers_ready_for_use/

Not sure whether to fo for the `analytics` cluster which is for
"batch jobs; long queries", or the `web` cluster for "real-time".
We probably don't need hard-core realtime so let's be on the safe side.

Bug: T176733

Change-Id: I28a4d15da823bbeaab81463c556d500b22c4d3ac