Ensure unicode encoding of query results
Switch from cl_sort_key to cl_sortkey_prefix as we are never
interested in the default (undefined) sort key. cl_sortkey_prefix
returns an empty string when no sortkey was set, this must be
handled so as not to crash splitlines().

As cl_sortkey and cl_sortkey_prefix may contain partial characters
(see T200623), any characters which cannot be decoded must be
replaced.

Bug: T200325
Change-Id: I207086bb03b948a4c3dbb7421d7fddcaf5732a70