Handle missing lat, lon in monument config
Lat and lon are no longer required in monuments config, thus the
related table keys should not always be created.

Add a check that if either lat or lon is provided then the other is
also provided.

A possible future imporvement would be allowing lat, lon to have
other names.

Bug: T176845
Change-Id: I3a5f00de3945b7a3ec19949ffca92d105d88a5ac