Fix logs path in Shell defaults to home directory
In 5510585 we moved all variable into one Shell file `defaults.sh`,
including a variable `LOGS_PATH`.

In 9742028 we made use of that variable for the first time,
for the execution of the categorisation job.

It turned out that this variable was wrongly defined in the first place:
the log directory is in the home directory, not in the source directory.

Effectively, `$SOURCE_PATH/logs` resolves to `/data/project/heritage/heritage/logs`.

Since the jsub execution was doing an append on the crontab file in that path,
it was failing with
`/data/project/heritage/heritage/logs/crontab.log: No such file or directory`

Bug: T202253
Change-Id: Iedd186108542b3629536ad80f1d9dd0abe4919fe