Ensure Python container waits for the database to be ready
Although docker-compose ensures that dependant containers
are started in the correct order, it does not ensure that
the underlying application is actually ready.

In our case, the Python application (aka bot) depends on
the database, which can take a while to boot as it has to
process all the SQL files generated by the Python.

This adds a Docker entrypoint which waits for the database
connection to succeed. This is done using the actual Python
application code, instead of doing it in Bash.

This has some advantages:
- no configuration/credentials to handle in the waiting script
- 1-1 parity with what the application actually needs

And some drawbacks:
- cannot be reused for the PHP container

Change-Id: I08a740a1f6e441d88225f375a69932ba732f911c