Make the admin structure tree update the last step
The udpate of the admin structure tree has a tendency
of crashing non-gracefully (because of eg memory issues).

This means that the rest of the daily update is not run,
which include statistics and other maintenance reports,
things that we care much more about than this one step.

This makes this step the last one of the process,
so that if it crashes, the other things are fine.

Change-Id: I00097bda12ab4c6e1ba97ebf79a26d6aea8a2b40