Base Python bot Dockerfile on ubuntu:14.04 instead of `python` image
The toolforge environment runs on Ubuntu 'Trusty' 14.04,
with Python 2.7.6. This should be mimiced as closely
as possible in the local development environment.

However, the `python:2.7` Docker image ships with the
latest minor version (at time of writing 2.7.15) and
older version are not available.

This changes the base image for Ubuntu 14.04,
which has python 2.7.6.

This would for example allow to reproduce locally T200159.

Change-Id: Iaed16d0c623acfd581dccc252e9f71313b8bd6d6