Install the latest Pip in the virtual environment
Since 4b9991d we are fully using a virtual environment
for everything Python related.

This environment by default has pip 1.5.4 (from 2014-02-21).

Such an old pip version is not compatible with our current
`requirements.txt` format.

This ensures we upgrade pip to the latest version
before using it to install packages.

Bug: T203432
Change-Id: I40e9db5eaf3045f898ddd3d8df588b6a119f2164