Add Bash script to run arbitrary Erfgoed Python script
The Python scripts cannot be in general invoked directly,
as one needs first to source the virtual environment
(which generally also means sourcing the vars in `defaults.sh`).

For that reason, we run ErfgoedBot processes using Bash scripts
such as `update_monuments.sh` and `categorize_images.sh`.

It however becomes a bit silly if we have to create a new Shell
script for every Python script, if there is no additional steps
to do (contrary to `update_monuments.sh`.

This adds a general purpose Bash script which takes as arguments
the Erfgoedbot Python script name and its arguments, and executes it.

An execution might look like:
```
./bin/run_erfgoedbot_script.sh erfgoedbot/update_database.py -countrycode:ge -langcode:ka -log
```

Will somehow help for T204036.

Change-Id: I835c5ae3fd3c0795c4f54564e62a9e09b54bf425