Switch API to use PHP 7.2
The Toolforge environment has been moving away from 5.6 to 7.2:
while 5.6 containers for webservices are still available,
the other containers are based on Debian Jessie and ship with 7.2.

This causes issues as we are running some maintenance scripts
from a non-PHP container (which we should not, really).

As upgrading to 7.2 is not too involved, let's just do that:
- update MySQL interface to use `mysqli`, as `mysql` was removed
- alter slightly the ResultsWrapper iterator, so that it still works
- update Dockerfile to use 7.2 (and also clean-up the PHP extensions
  installed, as most of them were not necessary).

Bug: T231223
Change-Id: Id33124d722e1eded2e25cadd7ae2e797c8daddb1