Remove the `security` extra for `requests` and `cryptography`
This was originall added in d2ff281 as workaround the outdated Python 2.7.6 from Toolforge.
As Toolforge is now running 2.7.19, this should not be necessary anymore.

This causes some changes in `Pipfile.lock` and requirements files,
insofar as `pipenv lock --keep-outdated` requires: -

- Removals: asn1crypto, cffi, enum34, ipaddress, pycparser, pyopenssl

In an upcoming changeset we will update the rest of the packages
(by not using `--keep-outdated` in pipenv)

Reverts: d2ff28187c64a2a3397781b0f913e69524509b84

Change-Id: I15638c38d49b5d96d3526677ec37593513f50abb