Surround all processCountry with try/except
We do not want to abort the all process just because
something went wrong with one country.
But we want to know wich couintry went wrong so we can investigate.

We had alraedy done that with update_database in c506d5f2,
and that should be generalized to all scripts.

Change-Id: I8fd1b065e57aa791448d52024cc75b8c7c908b55