Open/close database connection for every dataset in categorize_images
The categorization job is failing very often - at a point where
most datasets are never actually processed.

Digging into the logs, it appears that one dataset will fail with
`(2006, "MySQL server has gone away (error(32, 'Broken pipe'))")`,
and all subsequent datasets will instantly fail.

(This is what the `ping` is supposed to fix but clearly not).

Instead of opening the database connection at the beginning of the
script and holding it for the complete run, let's open it for every
dataset and close it once done.

Change-Id: I883f8d298e52e81f890ebf7be7b7515eb1b1c9d7