Switch to new Debian Buster toolforge Grid and Python3.7
The previous Tools Grid running on Strech is deprecated.
It was running Python 3.5.

The new Tools Grid running on Debian Buster ships with Python3.7.

We thus need to switch ErfgoedBot to use Python3.7
(which we were already testing for since 0955c336).

We also need to rebuild the virtual environment as Python3.7,
and doing so from a bastion host running Debian Buster. [1]

- Update all scaffolding (Pipfile, Pipfile.lock, tox file) to use Py3.7

- Base Dockerfile on `buster-slim` rather than stretch-slim (set in 7e78abb4)
  (and remove setuptools workaround added in f43b6aa2)

- Build virtual env with Python 3.7 instead of 3.5 (set in 0955c336)
  in python build Shell script

- Connect to a Buster Bastion host in SSH shell script

- Add `-release buster` flag to jsub command in update_monuments.sh

[1] See https://wikitech.wikimedia.org/wiki/News/Toolforge_Stretch_deprecation

Bug: T307269
Change-Id: I0c996462c04951e16a91197fef3de7fc2d5acf2c