Fix jsub syntax in update_monuments Shell script
The `update_monuments` Shell script has a step
which updates the `monuments_all` table.

This was often failing because the categorization job
was running in the background, preventing the update.
Since 974202827, the script stops the categorization job
before the update, and restarts it afterwards, using `jsub`

In 11fb1b0, we updated the call to `jsub` with `-release stretch`,
as part of the Toolforge Stretch deprecation [1].

We got the syntax wrong: the `-l release=trusty` was the syntax
recommended as part of the Toolforge Precise deprecatton [2].
We left over that spurious `-l` flag, and the call was failing with:

```
2022-09-07_13:30:22 Restart the categorization job...
usage: jsub [options...] program [args...]
jsub: error: argument -l: expected 1 argument
```

[1] https://wikitech.wikimedia.org/wiki/News/Toolforge_Stretch_deprecation
[2] https://wikitech.wikimedia.org/wiki/News/Tools_Precise_deprecation

Bug: T317278
Change-Id: I711807e738d70dc03ae9a217bf89271ac2688a2a