Remove `STRICT_TRANS_TABLES` SQL mode from our MariaDB database
Per T301949, ToolsDB was upgraded in April 2023
from MariaDB 10.1 (Debian Stretch) to 10.4 (Buster

With 10.4, the SQL model `STRICT_TRANS_TABLES` [1] was made default.

This results in harvesting crashing with `Data too long for column`.

The underlying issue is that data in the Wiki tables is 'invalid',
which we cannot really fix so easily.

Easiest path forward is to disable `STRICT_TRANS_TABLES` mode.

This was done directly in the ToolsDB database on Toolforge,
and here we reflect that in the docker-compose setup.

[1] https://mariadb.com/kb/en/sql-mode/#strict_trans_tables

Bug: T338987
Change-Id: I59d22372b433d6571d3aaa6471195af1a95218ec