Replace deprecated `minorEdit` argument with `minor`
pywikibot has deprecated the `minorEdit` argument
of `put` in favour of `minor`, resulting in many
warnings in the logs [1].

Formally we only need to replace the call to `page.put`
in `common.py`, but let's replace everywhere for consistency.

[1] Message:
```
FutureWarning: minorEdit argument of pywikibot.page.BasePage.put is deprecated; use minor instead.
```

Change-Id: Ic7cac6e39034837cac5e8e1670cdb1c194364f7c