Do not use `STRICT_TRANS_TABLES` SQL mode when connecting to database
Per T301949, ToolsDB was upgraded in April 2023
from MariaDB 10.1 (Debian Stretch) to 10.4 (Buster

With 10.4, the SQL model `STRICT_TRANS_TABLES` [1] was made default.

This results in harvesting crashing with `Data too long for column`.

The underlying issue is that data in the Wiki tables is 'invalid',
which we cannot really fix so easily.

Easiest path forward is to disable `STRICT_TRANS_TABLES` mode.
We had tried to do that directly in the ToolsDB database on Toolforge,
using a `set` command, but that does not persist.

Here instead we set it in `pymysql.connect`.

(This was already reflected in the docker-compose setup in 8e60754e06).

[1] https://mariadb.com/kb/en/sql-mode/#strict_trans_tables

Bug: T338987
Change-Id: I235c2a95d7f26893e8c15d69dcebc40d12a644b7