Set file permissions on database configuration file in Docker build
When developing locally, I had issues with permissions on
files mounted via volumes (which is solved by a local chmod),
but also on that file which is copied over in the Dockerfile.

The `chmod` flag to Docker's COPY command has been available
for a while now, so let's use it.

See also https://docs.docker.com/reference/dockerfile/#copy---chown---chmod

Change-Id: I15d96f29310a08208907ec4905f8fa3a87567046