Use toolforge-jobs to install requirements during deployment
As the Toolforge GridEngine has been deprecated,
we need to migate the cron jobs to its replacement
Toolforge Kubernetes.

We have Python requirements, which must
be installed in a virtual environment.
Per the Toolforge documentation:
```
A virtual environment must be created in the same
execution environment that the venv will be used from.
This means that any virtual enviroments must be
bootstrapped inside a container using a job.
```

- Update `build-python.sh` following the Toolforge recommendations
  (as enshrined in the `webservice-python-bootstrap` script)

- Invoke `build-python.sh` and `build-php.sh` directly from `deploy.sh`,
  using toolforge-jobs, which requires making the files executable

- Remove `build.sh` as unused

(Per the Toolforge docs, it was probably not necessary to
run `build-php` from a job, but let's do it for consistency).

https://wikitech.wikimedia.org/wiki/Help:Toolforge/Python#Virtual_environments_with_prebuilt_images
https://phabricator.wikimedia.org/diffusion/ODIT/browse/master/shared/python/webservice-python-bootstrap
https://wikitech.wikimedia.org/wiki/Help:Toolforge/PHP

Bug: T319787
Change-Id: I1be51fa0962e3ec47eec20cf87ca93c937f7d23c