Rewrite jobs shell scripts for Toolforge Jobs framework
The jobs that used to run on the now-deprecated GridEngine
need to be changed for the Toolforge Jobs framework.

The biggest changes are:

- Jobs run in a particular container (python, php...)
  so all necessary runtimes are not available at once.

- the harvesting job used to pause (stop/start) the
  categorization job, which is not possible anymore.

- Jobs can start other jobs over the Toolforge API

Here we split the `update_monuments` in several chunks:

- `harvesting.sh` (Python container) runs the harvesting
  and triggers `fill_table_monuments_all.sh`

- `fill_table_monuments_all.sh` (mariadb container)
  runs the SQL merge, and triggers 3 jobs:

  - `post_harvesting_php.sh` (PHP container) takes care of
    detailed statistics, prox_search table and admin structure tree

  - `post_harvesting_python.sh` (Python container) does:
    statistics, unused images list, list of images without id,
    list of missing commonscat links.

  - `categorize_images.sh` (Python container),
    which categorizes pictures and which we now
    expect to complete before next harvesting
    (removing the need for stop/start)

The Python jobs are changed so that the source code
is added to `PYTHONPATH` ; otherwise Python was
complaining it could not find the modules.

There were some other steps in `update_monuments.sh`,
that have not been moved here for simplicity:
- table recreation, since it uses  two runtimes
  (Python to create the SQL, and mariadb to run the SQL)
- database dump, as I’m not sure anyone uses that
  (but could probably be added as sub-job)

(Also, arguably, that step should be done as part of
deploying the code rather than at each nightly run).

Useful links about the Toolforge API:

https://wikitech.wikimedia.org/wiki/Help:Toolforge/API
https://api-docs.toolforge.org/docs

Bug: T319787
Change-Id: Ie2f98e1fe28b3f796c72ae4447ac0848a0741edd